Extract shared message rendering functions into message-renderers.ts
Problem: renderSubagentCard, renderToolCallBlock, and the assistant
message child-rendering logic were duplicated between MessageList.ts
and SubagentView.ts, creating maintenance burden and divergence risk.

Fix: Extract the shared rendering functions into a new
message-renderers.ts module and import them from both consumers.
